% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_expedition.R
\name{create_expedition_readme}
\alias{create_expedition_readme}
\title{Create Expedition README Template}
\usage{
create_expedition_readme(expedition_name)
}
\description{
Create Expedition README Template
}
\keyword{internal}
