% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create New Pristine Seas Project Structure}
\usage{
create_project(project_name)
}
\arguments{
\item{project_name}{Character string. Name of the new project (use "prj-" prefix by convention)}
}
\value{
Character string. Path to the created project folder
}
\description{
Creates a simple standardized folder structure for Pristine Seas projects in the
projects folder of the science drive.
}
\examples{
\dontrun{
# Create a project
create_project("prj-coral-reef-assessment")

# Create another project
create_project("prj-fish-surveys")
}
}
