% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PristineSeasR2-package.R
\docType{package}
\name{pristineseas-package}
\alias{pristineseas}
\alias{pristineseas-package}
\title{pristineseas: Marine Conservation Data Analysis Tools}
\description{
Tools for analyzing marine biodiversity data from Pristine Seas expeditions. Includes functions for taxonomic data cleaning and spatial analysis.
}
\author{
\strong{Maintainer}: Juan Mayorga \email{marine.data.science@ngs.org}

}
\keyword{internal}
