% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_researchers.R
\docType{data}
\name{ps_researchers}
\alias{ps_researchers}
\title{Scientists Lookup Table}
\format{
A data frame with scientist information
}
\usage{
ps_researchers
}
\description{
Directory of Pristine Seas scientists and collaborators with contact information,
affiliations, and research specialties.
}
\details{
\strong{status} categories:
\itemize{
\item \strong{core} — Full-time Pristine Seas staff
\item \strong{affiliate} — Regular collaborators and partners
\item \strong{visiting} — Visiting scientists and temporary collaborators
\item \strong{Emeritus} - Retired team member
\item \strong{Local} - Local participant
}
}
\keyword{datasets}
