% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PristineSeasR2-package.R
\docType{package}
\name{PristineSeasR2-package}
\alias{PristineSeasR2}
\alias{PristineSeasR2-package}
\title{PristineSeasR2: Tools for Pristine Seas Scientific Data Analysis and Visualization}
\description{
Provides helper functions for analyzing Pristine Seas scientific data, custom color palettes and themes for data visualization, and standardized vocabularies for consistent data analysis across expeditions. Designed for internal use by Pristine Seas research team to enhance workflow efficiency and maintain analytical consistency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pristine-seas/PristineSeasR2}
  \item \url{https://pristine-seas.github.io/PristineSeasR2/}
  \item Report bugs at \url{https://github.com/pristine-seas/PristineSeasR2/issues}
}

}
\author{
\strong{Maintainer}: Juan Mayorga \email{marine.data.science@ngs.org}

}
\keyword{internal}
