% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PristineSeasR2-package.R
\docType{package}
\name{PristineSeasR2-package}
\alias{PristineSeasR2}
\alias{PristineSeasR2-package}
\title{PristineSeasR2: Marine Conservation Data Analysis Tools for Pristine Seas}
\description{
Tools for analyzing marine biodiversity data from Pristine Seas expeditions. Includes functions for taxonomic data cleaning, standardized vocabularies, color palettes for visualization, and spatial analysis utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pristine-seas/PristineSeasR2}
  \item \url{https://pristine-seas.github.io/PristineSeasR2/}
  \item Report bugs at \url{https://github.com/pristine-seas/PristineSeasR2/issues}
}

}
\author{
\strong{Maintainer}: Juan Mayorga \email{marine.data.science@ngs.org}

}
\keyword{internal}
