% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_colors.R
\name{ps_colors}
\alias{ps_colors}
\title{Get Pristine Seas Colors}
\usage{
ps_colors(palette)
}
\arguments{
\item{palette}{Character. Palette name}
}
\value{
Named vector of hex color codes
}
\description{
Get Pristine Seas Colors
}
\examples{
\dontrun{
ps_colors("trophic_group")
ps_colors("functional_groups")

# Use in ggplot2
scale_color_manual(values = ps_colors("trophic_group"))
}
}
