% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ps.R
\name{theme_ps}
\alias{theme_ps}
\title{Pristine Seas ggplot2 theme}
\usage{
theme_ps(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{Numeric. Base font size. Default is 12.}

\item{base_family}{Character. Base font family. Default is "Helvetica".}
}
\value{
A ggplot2 theme object.
}
\description{
A clean, publication-ready ggplot2 theme
}
\examples{
library(ggplot2)
ggplot(mpg, aes(class, hwy)) +
  geom_boxplot(fill = "#0A9396", color = "white") +
  labs(title = "Fuel efficiency by car class") +
  theme_ps()

}
